Fixed a divide by zero exception, now the dividion is rounded to two decimal places, and user are now able to enter decimal answer without errors.

+++ MathWizzz/MathWizzz/Models/Drill.cs
@@ -47,7 +47,7 @@ namespace MathWizzz
             }
         }
 
-        public override bool CheckAnswer(int studentAnswer)
+        public override bool CheckAnswer(double studentAnswer)
         {
             bool isCorrect = Question.CheckAnswer(studentAnswer);
